// src/generated-api/models/Effects.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Underworld
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.6.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    Attack,
    AttackFromJSON,
    AttackFromJSONTyped,
    AttackToJSON,
} from './Attack';
import {
    Defense,
    DefenseFromJSON,
    DefenseFromJSONTyped,
    DefenseToJSON,
} from './Defense';
import {
    Poison,
    PoisonFromJSON,
    PoisonFromJSONTyped,
    PoisonToJSON,
} from './Poison';

/**
 * 
 * @export
 * @interface Effects
 */
export interface Effects {
    /**
     * 
     * @type {Defense}
     * @memberof Effects
     */
    shield_aura?: Defense;
    /**
     * 
     * @type {boolean}
     * @memberof Effects
     */
    knows_has_shield_aura: boolean;
    /**
     * 
     * @type {Attack}
     * @memberof Effects
     */
    retribution_aura?: Attack;
    /**
     * 
     * @type {boolean}
     * @memberof Effects
     */
    knows_has_retribution_aura: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Effects
     */
    resurrection_aura: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Effects
     */
    knows_has_resurrection_aura: boolean;
    /**
     * 
     * @type {Poison}
     * @memberof Effects
     */
    poison?: Poison;
    /**
     * 
     * @type {boolean}
     * @memberof Effects
     */
    covered_in_oil: boolean;
}

export function EffectsFromJSON(json: any): Effects {
    return EffectsFromJSONTyped(json, false);
}

export function EffectsFromJSONTyped(json: any, ignoreDiscriminator: boolean): Effects {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'shield_aura': !exists(json, 'shield_aura') ? undefined : DefenseFromJSON(json['shield_aura']),
        'knows_has_shield_aura': json['knows_has_shield_aura'],
        'retribution_aura': !exists(json, 'retribution_aura') ? undefined : AttackFromJSON(json['retribution_aura']),
        'knows_has_retribution_aura': json['knows_has_retribution_aura'],
        'resurrection_aura': json['resurrection_aura'],
        'knows_has_resurrection_aura': json['knows_has_resurrection_aura'],
        'poison': !exists(json, 'poison') ? undefined : PoisonFromJSON(json['poison']),
        'covered_in_oil': json['covered_in_oil'],
    };
}

export function EffectsToJSON(value?: Effects | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'shield_aura': DefenseToJSON(value.shield_aura),
        'knows_has_shield_aura': value.knows_has_shield_aura,
        'retribution_aura': AttackToJSON(value.retribution_aura),
        'knows_has_retribution_aura': value.knows_has_retribution_aura,
        'resurrection_aura': value.resurrection_aura,
        'knows_has_resurrection_aura': value.knows_has_resurrection_aura,
        'poison': PoisonToJSON(value.poison),
        'covered_in_oil': value.covered_in_oil,
    };
}